# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Raumdeuter API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/players/{wyscout_id}/xtv-history": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Xtv History",
        "description": "Up to ``limit`` consecutive loaded seasons ending at ``season``, non-null xTV rows only (chronological).",
        "operationId": "xtv_history_players__wyscout_id__xtv_history_get",
        "parameters": [
          {
            "name": "wyscout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Wyscout Id"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Anchor season (inclusive) \u2014 last N loaded seasons up to this year.",
              "title": "Season"
            },
            "description": "Anchor season (inclusive) \u2014 last N loaded seasons up to this year."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10,
              "minimum": 1,
              "default": 5,
              "title": "Limit"
            }
          },
          {
            "name": "club",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "When set, prefer this club's row each season (fallback: most minutes).",
              "title": "Club"
            },
            "description": "When set, prefer this club's row each season (fallback: most minutes)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XtvHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/players/{wyscout_id}/performance-index-history": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Performance Index History",
        "description": "Up to ``limit`` consecutive loaded seasons ending at ``season``, non-null PI rows only (chronological).\n\nOne row per season: ``club`` when provided, else the club with most minutes (profile default).",
        "operationId": "performance_index_history_players__wyscout_id__performance_index_history_get",
        "parameters": [
          {
            "name": "wyscout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Wyscout Id"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Anchor season (inclusive) \u2014 last N loaded seasons up to this year.",
              "title": "Season"
            },
            "description": "Anchor season (inclusive) \u2014 last N loaded seasons up to this year."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10,
              "minimum": 1,
              "default": 5,
              "title": "Limit"
            }
          },
          {
            "name": "club",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "When set, prefer this club's row each season (fallback: most minutes).",
              "title": "Club"
            },
            "description": "When set, prefer this club's row each season (fallback: most minutes)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceIndexHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/players/{wyscout_id}/profile": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Profile",
        "operationId": "profile_players__wyscout_id__profile_get",
        "parameters": [
          {
            "name": "wyscout_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Wyscout Id"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Season"
            }
          },
          {
            "name": "club",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Disambiguates when a player has rows for multiple clubs in the same season.",
              "title": "Club"
            },
            "description": "Disambiguates when a player has rows for multiple clubs in the same season."
          },
          {
            "name": "league",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Percentile cohort league; omit to use this player's league",
              "title": "League"
            },
            "description": "Percentile cohort league; omit to use this player's league"
          },
          {
            "name": "min_minutes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 200,
              "title": "Min Minutes"
            }
          },
          {
            "name": "performance_index_history_limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 10,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "When set, include up to N PI trajectory seasons inside this response.",
              "title": "Performance Index History Limit"
            },
            "description": "When set, include up to N PI trajectory seasons inside this response."
          },
          {
            "name": "x_tv_history_limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 10,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "When set, include up to N xTV trajectory seasons inside this response.",
              "title": "X Tv History Limit"
            },
            "description": "When set, include up to N xTV trajectory seasons inside this response."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PerformanceIndexHistoryPoint": {
        "properties": {
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "performance_index": {
            "type": "number",
            "title": "Performance Index"
          },
          "club": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Club"
          },
          "club_logo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Club Logo"
          },
          "minutes_played": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minutes Played",
            "description": "Season minutes for this row; UI hides PI headline under typical radar threshold."
          }
        },
        "type": "object",
        "required": [
          "season",
          "performance_index"
        ],
        "title": "PerformanceIndexHistoryPoint",
        "description": "Single season snapshot for scout profile PI progression (club = dominant minutes row)."
      },
      "PerformanceIndexHistoryResponse": {
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/PerformanceIndexHistoryPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "points"
        ],
        "title": "PerformanceIndexHistoryResponse",
        "description": "Chronological (oldest \u2192 newest), at most ``limit`` seasons with non-null PI."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "XtvHistoryPoint": {
        "properties": {
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "x_tv_eur": {
            "type": "number",
            "title": "X Tv Eur"
          },
          "club": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Club"
          },
          "club_logo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Club Logo"
          }
        },
        "type": "object",
        "required": [
          "season",
          "x_tv_eur"
        ],
        "title": "XtvHistoryPoint",
        "description": "Single season snapshot for scout profile xTV progression (club = dominant minutes row)."
      },
      "XtvHistoryResponse": {
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/XtvHistoryPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "points"
        ],
        "title": "XtvHistoryResponse",
        "description": "Chronological (oldest \u2192 newest), at most ``limit`` seasons with non-null xTV."
      },
...
    }
  }
}